Huge changes related to Parsing Immediate Instructions
Fixed many tests, NEED ERROR REPORTER TO BE FULLY RESTRUCTURED!!!

+++ IMnemonic.java
@@ -3,7 +3,7 @@
  */
 public interface IMnemonic extends IToken{
 
-    byte getOpcode();
+    int getOpcode();
     boolean isValidOperation();
     String getName();
     void setName(String name);
